backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.8.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
alembic==1.15.1
httpx==0.28.1
itsdangerous==2.2.0
aiofiles==24.1.0
boto3==1.37.8
pyotp==2.9.0
slowapi==0.1.9